Added the part of the Quiz that makes use of checkboxes
Per reviewers suggestion, I created a graphical/diagram implementation of checkboxes, which I called create-a-chord. To get the right answer, the correct 3 checkboxes should be toggled on. This represents the strings/fret fingering guitar players left hand activity in playing the guitar. So, 3/24 is the 1 correct answer. I can later change this to accommodate other chords. For now an open G is created by toggling on checkboxes with id: fret3string1, fret2string2, and fret3string6.